scripts/requirements.txt
transformers>=4.35.0
pytorch-transformers>=1.2.0
accelerate>=0.24.0